ChangesMadeOnDeleteFunctions

+++ src/main/java/com/example/model/Student.java
@@ -1,22 +1,21 @@
 package com.example.model;
 
 import javax.persistence.*;
-import java.util.Set;
 
 /**
  * Parent Table =  Student (Contains the Primary Key)
  * Child Table = Registration (Contains the Foreign key of the Parent class)
  * One To Many = One entry of the Parent class can be used in many/multiple
- *              entries of the Child class.
+ * entries of the Child class.
  * Mapped by = The mappedBy property is what we use to tell Hibernate which
- *              variable we are using to represent the parent class in our child class.
+ * variable we are using to represent the parent class in our child class.
  */
 
 @Entity
 public class Student {
 
     @Id
-    @GeneratedValue(strategy= GenerationType.SEQUENCE)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
     int studentId;
     @Column(name = "fname")
     String firstName;
@@ -25,15 +24,12 @@ public class Student {
     @Column(name = "age")
     int age;
     @Column(name = "contactno")
-    int contactNo;
+    long contactNo;
 
     public Student() {
     }
 
-
-
-
-    public Student(int studentId, String firstName, String lastName, int age, int contactNo) {
+    public Student(int studentId, String firstName, String lastName, int age, long contactNo) {
         this.studentId = studentId;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -73,19 +69,12 @@ public class Student {
         this.age = age;
     }
 
-    public int getContactNo() {
+    public long getContactNo() {
         return contactNo;
     }
 
-    public void setContactNo(int contactNo) {
+    public void setContactNo(long contactNo) {
         this.contactNo = contactNo;
     }
 
-//    public Set<Registration> getRegistrations() {
-//        return registrations;
-//    }
-//
-//    public void setRegistrations(Set<Registration> registrations) {
-//        this.registrations = registrations;
-//    }
 }
